# openapi.yaml
openapi: 3.0.0
info:
  title: NjordBreeze API
  version: 0.1.0
  description: >-
    API for integrating personal weather station data with SMHI open data for
    comparison and analysis.
  contact:
    name: Fredrik Svensson
    email: fs222id@student.lnu.se
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://cscloud8-57.lnu.se/njordbreeze/
    description: Production server
paths:
...
  /users/{id}:
    get:
      summary: Get information about a specific user
      description: Returns information about a specific user in the system.
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID.
      responses:
        '200':
          description: Successful response with user information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found.
    put:
      summary: Update information about a specific user
      description: Updates information about a specific user in the system.
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '200':
          description: User updated successfully.
        '404':
          description: User not found.
    delete:
      summary: Delete a specific user
      description: Deletes a specific user from the system.
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The user's ID.
      responses:
        '200':
          description: User deleted successfully.
        '404':
          description: User not found.
...
components:
  schemas:
    NewUser:
      type: object
      properties:
        name:
          type: string
          description: The user's full name.
        email:
          type: string
          format: email
          description: The user's email address.
...
    User:
      type: object
      properties:
        id:
          type: string
          description: The user's ID.
        name:
          type: string
          description: The user's full name.
        email:
          type: string
          format: email
          description: The user's email address.
...